Get flag via contract market info
