+++ src/components/member/tabs/quote-tab/create-quote-form.tsx
@@ -7,7 +7,7 @@ import { JsonSchemaForm } from 'hedvig-ui/json-schema-form'
 import React, { useState } from 'react'
 import { Checkbox } from 'semantic-ui-react'
 import { WithShowNotification } from 'store/actions/notificationsActions'
-import { ContractType } from 'utils/contract'
+import { ContractType } from 'types/enums'
 import { withShowNotification } from 'utils/notifications'
 
 const CreateQuoteFormComponent: React.FC<{
