+++ src/components/member/tabs/quote-tab/quote-sub-section.tsx
@@ -6,7 +6,7 @@ import { MainHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import styled from 'react-emotion'
 import { showNotification } from 'store/actions/notificationsActions'
-import { ContractType } from 'utils/contract'
+import { ContractType } from 'types/enums'
 import { isSigned, isSignedOrExpired } from 'utils/quote'
 import { ActionsWrapper, Muted } from './common'
 import { QuoteListItem } from './quote-list-item'
