+++ src/types/enums.ts
@@ -12,3 +12,13 @@ export enum QuoteProductType {
   Travel = 'TRAVEL',
   Accident = 'ACCIDENT',
 }
+
+export enum ContractType {
+  SwedishApartment = 'SWEDISH_APARTMENT',
+  SwedishHouse = 'SWEDISH_HOUSE',
+  NorwegianHomeContent = 'NORWEGIAN_HOME_CONTENT',
+  NorwegianTravel = 'NORWEGIAN_TRAVEL',
+  DanishHomeContent = 'DANISH_HOME_CONTENT',
+  DanishTravel = 'DANISH_TRAVEL',
+  DanishAccident = 'DANISH_ACCIDENT',
+}
